+++ src/components/Common/Toast.jsx
@@ -1,8 +1,12 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import styled, { css, keyframes } from 'styled-components';
 import Completed from '../../assets/icons/completed.svg';
 import Close from '../../assets/icons/close.svg';
+import {
+  ContainerDiv,
+  ToastBoxDiv,
+  ToastCompleted,
+} from '../../styles/Common/Toast.style';
 
 function Toast({ showToast, setShowToast }) {
   // Toast 5초뒤에 꺼지는 코드
@@ -42,63 +46,3 @@ Toast.propTypes = {
 };
 
 export default Toast;
-
-const ContainerDiv = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0 24px;
-`;
-
-const timer = keyframes`
-  from{
-    width: 100%;
-  }to{
-    width: 0%;
-  }
-`;
-
-const timerAnimation = css`
-  animation: ${timer} 5s linear;
-`;
-
-const ToastBoxDiv = styled.div`
-  position: fixed;
-  bottom: 0px;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  width: calc(100vw - 48px);
-  max-width: 524px;
-  height: 64px;
-  padding: 19px 30px;
-  border-radius: 8px;
-  background: rgba(0, 0, 0, 0.8);
-  opacity: ${({ $showToast }) => ($showToast ? '1' : '0')};
-  transform: translateY(${({ $showToast }) => ($showToast ? '-70px' : '0px')});
-  transition: all 0.5s cubic-bezier(0.68, -0.55, 0.25, 1.35);
-  &::before {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    height: 5px;
-    width: 0;
-    background-color: #40f467;
-    border-radius: 0 0 8px 8px;
-    ${({ $showToast }) => ($showToast ? timerAnimation : '')};
-  }
-`;
-
-const ToastCompleted = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  .CompletedText {
-    color: var(--white, #fff);
-    font-size: var(--font16, 1.6rem);
-    font-weight: var(--regular, 400);
-    line-height: 26px;
-    letter-spacing: -0.16px;
-  }
-`;
