+++ src/components/RollingBgPage/BgList.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import BgImg from './BgImg';
 import BgColor from './BgColor';
+import { BgListWrapper } from '../../styles/RollingToBgPage/BgList.style';
 
 function BgList({ images, onSelectColor, onSelectImg, selectedColor }) {
   const [selectedImg, setSelectedImg] = useState('');
@@ -55,18 +55,3 @@ BgList.defaultProps = {
 };
 
 export default BgList;
-
-const BgListWrapper = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 100%;
-  margin-bottom: 69px;
-  @media (min-width: 360px) and (max-width: 767px) {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 12px;
-    margin-bottom: 58px;
-  }
-`;
