+++ src/components/RollingPage/AddCard.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PropTypes } from 'prop-types';
-import { styled } from 'styled-components';
 import { ReactComponent as Plus } from '../../assets/icons/plus.svg';
+import { AddCardDiv, PlusFrame } from '../../styles/RollingPage/AddCard.style';
 
 function AddCard({ onClick }) {
   return (
@@ -20,35 +20,5 @@ AddCard.propTypes = {
 AddCard.defaultProps = {
   onClick: null,
 };
-
-const AddCardDiv = styled.div`
-  width: 100%;
-  min-height: 280px;
-  height: 100%;
-  padding: 0px 24px;
-  border-radius: 16px;
-  background: ${({ theme }) => theme.btnBgColor};
-  box-shadow: 0px 2px 12px 0px rgba(0, 0, 0, 0.08);
-  cursor: pointer;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const PlusFrame = styled.div`
-  width: 56px;
-  height: 56px;
-  border-radius: 100px;
-  background-color: var(--gray500);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: auto;
-
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
 
 export default AddCard;
