+++ src/pages/ListPage.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import styled from 'styled-components';
 import ListCard from '../components/ListPage/ListCard';
-import LinkButton from '../components/LandingPage/Button';
 import Header from '../components/Common/Header/Header';
+import { Container, Wrapper } from '../styles/ListPage/ListPage.style';
+import { LinkButton } from '../styles/LandingPage/LinkButton.style';
 
 function ListPage() {
   return (
@@ -19,23 +19,3 @@ function ListPage() {
 }
 
 export default ListPage;
-
-const Container = styled.main`
-  width: 100%;
-  min-height: calc(100vh - 65px);
-  display: flex;
-  justify-content: center;
-  background-color: ${({ theme }) => theme.bgColor};
-`;
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 50px;
-  margin-top: 50px;
-  padding: 0 24px;
-  @media all and (max-width: 1248px) {
-    overflow: hidden;
-  }
-`;
