+++ src/styles/LandingPage/LandingCard.style.js
