+++ src/styles/ListPage/CardData.style.js
