+++ src/styles/RollingPage/DetailCard.style.js
